доработка popup
