+++ js/main.js
@@ -1,4 +1,2 @@
 import { renderThumbnails } from './render-picture-thumbnails.js';
-import './pop-up.js';
-
-window.log (renderThumbnails);
+import './popup.js';
