+++ js/render-big-picture.js
@@ -1,4 +1,5 @@
 import { mockPhotos } from './data.js';
+import { bigPictureCommentsGenerator } from './comments-generator.js';
 
 const bigPicture = document.querySelector('.big-picture');
 const bigPictureImg = bigPicture.querySelector('img');
@@ -19,21 +20,6 @@ const renderBigPicture = (thumbnailId) => {
   if (photoID) {
     const { url, description, comments, likes } = photoID;
     createBigPictureTemplate(url, description, comments, likes);
-
-    // Генератор комментариев
-    const bigPictureCommentsGenerator = () => {
-      const commentsContainer = bigPicture.querySelector('.social__comments');
-      const singleCommentTemplate = commentsContainer.querySelector('.social__comment');
-      commentsContainer.innerHTML = '';
-
-      comments.forEach((comment) => {
-        const commentElement = singleCommentTemplate.cloneNode(true);
-        commentElement.querySelector('.social__picture').src = comment.avatar;
-        commentElement.querySelector('.social__picture').alt = comment.name;
-        commentElement.querySelector('.social__text').textContent = comment.message;
-        commentsContainer.append(commentElement);
-      });
-    };
 
     // Вызов функции генерации комментариев
     bigPictureCommentsGenerator(comments);
